Add Remark Command in AB3
The AB3 application doesn't have a feature to optionally add
remarks for the people in the address book. Having such a feature
would be useful so that more information can be known about the person.

Let's create and implement a remark command where a person in the
address book can have an optional remark.